backend/create/requirements.txt
Flask
Flask-CORS
requests
PyJWT